Rename stopper package name
