+++ pkg/common/stopper/stopper.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package stop
+package stopper
 
 import (
 	"context"
